requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
pydantic
firebase-admin
